requirements.txt
# YAPAY ZEKÂ DESTEKLİ PROFESYONEL ÇAĞRI MERKEZİ
# Gereksinimler / Requirements

# Web Framework
flask==3.0.0
flask-login==0.6.3
flask-wtf==1.2.1
flask-cors==4.0.0
flask-socketio==5.3.6

# Database
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
werkzeug==3.0.1
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# VoIP / WebRTC
python-sipsimple==5.0.0
twilio==8.10.0
pjsua==2.14
aiortc==1.6.0

# AI & Speech Processing
openai==1.6.1
whisper==1.1.10
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
langchain==0.0.352
langchain-openai==0.0.2

# Speech & NLP
speechrecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.14
webrtcvad==2.0.10
nltk==3.8.1
spacy==3.7.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Async & Real-time
eventlet==0.34.2
gevent==23.9.1
celery==5.3.4
kombu==5.3.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2

